pyproject.toml
[project]
name = "opensource-scout"
version = "0.1.0"
description = "Multi-agent GitHub automation for discovering and contributing to open-source projects"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "openai>=1.40.0"
]

[project.scripts]
scout = "apps.cli.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["packages*", "apps*"]
